Update better modularize for mnist example
